The directory tree has been organized
