+++ src/components/ContactButton/ContactButton.component.tsx
@@ -1,8 +1,8 @@
 import { useContext } from "react";
 import "firebase/compat/auth";
 import { useNavigate } from "react-router-dom";
-import { ButtonS } from "../components/Buttons/Button.styled";
-import { AuthContext, AuthContextType } from "../Context/AuthContext";
+import { ButtonS } from "../Buttons/Button.styled";
+import { AuthContext, AuthContextType } from "../../Context/AuthContext";
 import firebase from "firebase/compat/app";
 import "firebase/compat/auth";
