+++ src/components/Layout/Layout.component.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from "react";
-import Header from "../Header/Header";
-import { FooterLabel } from "../../Footer/Footer.component";
+import Header from "../Header/Header.component";
+import { FooterLabel } from "../Footer/Footer.component";
 import { Main } from "./Layout.styled";
 
 interface LayoutProps {
