+++ src/components/StartRentingButton/StartRentingButton.component.tsx
@@ -1,6 +1,6 @@
 import { useContext } from "react";
-import { AuthContext } from "../Context/AuthContext";
-import { ButtonM } from "../components/Buttons/Button.styled";
+import { AuthContext } from "../../Context/AuthContext";
+import { ButtonM } from "../Buttons/Button.styled";
 import "firebase/compat/auth";
 import { Link, useNavigate } from "react-router-dom";
