+++ src/components/UserPanelButton/UserPanelButton.component.tsx
@@ -1,4 +1,4 @@
-import { ButtonS } from "../components/Buttons/Button.styled";
+import { ButtonS } from "../Buttons/Button.styled";
 import { Link } from "react-router-dom";
 import React from "react";
