+++ src/pages/AddProductPage.tsx
@@ -1,4 +1,4 @@
-import CategoryDropdownSecond from "../components/CategoryDropdown/CategoryDropdownSecond";
+import CategoryDropdownSecond from "../components/CategoryDropdown/CategoryDropdownSecond.component";
 import { CategotyContainer } from "../components/AddProductPage/AddNewProduct.styled";
 
 export const AddProductPage = () => {
