+++ src/pages/HowDoesItWorkBorrowPage.tsx
@@ -3,7 +3,7 @@ import { ButtonContainer, ButtonM, OrangeButton } from "../components/Buttons/Bu
 import { HowDoesItWorkBorrow } from "../components/HowDoesItWork/HowDoesItWork.component";
 import { ButtonStartRenting, HowWork } from "../components/HowDoesItWork/HowDoesItWork.styled";
 import { Title } from "../UI/Title.styled";
-import { StartRentingButton } from "../StartRentingButton/StartRentingButton";
+import { StartRentingButton } from "../components/StartRentingButton/StartRentingButton.component";
 
 export const HowDoesItWorkBorrowPage = () => {
   return (
